Move logger to a logger package inside merlin.
This should prevent conflicts with other NOS projects that expose a support package.
