+++ core/src/main/java/com/novoda/merlin/registerable/connection/ConnectCallbackManager.java
@@ -2,7 +2,7 @@ package com.novoda.merlin.registerable.connection;
 
 import com.novoda.merlin.registerable.MerlinCallbackManager;
 import com.novoda.merlin.registerable.Register;
-import com.novoda.support.Logger;
+import com.novoda.merlin.logger.Logger;
 
 public class ConnectCallbackManager extends MerlinCallbackManager<Connectable> {
